- removed command_if class: not usefull, use python instead for complex logic - renamed item.delivery field to item.operation - minor bug fix in test script: file comparison didn't work fine on empty lines - renamed unit test calendar to constraints_resource_5 - added support for renaming HasName entities - calendar buckets didn't correctly initialize their value upon creation - additional constructor for XMLelement

+++ src/model/resource.cpp
@@ -113,7 +113,7 @@ DECLARE_EXPORT void Resource::writeElement(XMLOutput *o, const XMLtag& tag, mode
         o->BeginObject(Tags::tag_load_plan);
         o->writeElement(Tags::tag_date, lp->getDate());
         o->writeElement(Tags::tag_quantity, lp->getQuantity());
-        o->writeElement(Tags::tag_onhand, lp->getOnhand());
+        o->writeElement(Tags::tag_onhand, static_cast<float>(lp->getOnhand()));
         o->writeElement(Tags::tag_minimum, lp->getMin());
         o->writeElement(Tags::tag_maximum, lp->getMax());
         o->writeElement(Tags::tag_operation_plan, &*(lp->getOperationPlan()), FULL);
